cluster: use fqdn as url
