+++ pkg/util/etcdutil/etcdutil.go
@@ -15,19 +15,13 @@
 package etcdutil
 
 import (
-	"errors"
 	"fmt"
-	"net/url"
-	"regexp"
-	"strings"
 
 	"github.com/coreos/etcd-operator/pkg/util/constants"
 	"github.com/coreos/etcd/clientv3"
 
 	"golang.org/x/net/context"
 )
-
-var validPeerURL = regexp.MustCompile(`^\w+:\/\/[\w\.\-]+(:\d+)?$`)
 
 func ListMembers(endpoints []string) (*clientv3.MemberListResponse, error) {
 	cfg := clientv3.Config{
@@ -61,18 +55,6 @@ func RemoveMember(clientURLs []string, id uint64) error {
 	_, err = etcdcli.Cluster.MemberRemove(ctx, id)
 	cancel()
 	return err
-}
-
-func MemberNameFromPeerURL(pu string) (string, error) {
-	// url.Parse has very loose validation. We do our own validation.
-	if !validPeerURL.MatchString(pu) {
-		return "", errors.New("invalid PeerURL format")
-	}
-	u, err := url.Parse(pu)
-	if err != nil {
-		return "", err
-	}
-	return strings.Split(u.Host, ":")[0], nil
 }
 
 func CheckHealth(url string) (bool, error) {
